requirements.txt


# --- Web framework ---
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# --- Data collection ---
requests>=2.32.0
yfinance>=0.2.40
pydantic>=2.7.0

# --- Data processing ---
pandas>=2.2.0
pyarrow>=16.0.0          # parquet support for pandas

# --- Sentiment (used in next module) ---
transformers>=4.41.0
torch>=2.3.0
anthropic>=0.28.0        # Claude API for LLM-based scoring

# --- Testing ---
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0            # for FastAPI TestClient

# --- Dev tools ---
python-dotenv>=1.0.0     # load .env files locally